remove catch-all handler, manually handle sub-view linking
Manually adds NavigateTo Msg variants for subviews that need to change
the url of the page, that way the updateUrl handler can listen for the
update and re-route accordingly. Unfortunately there is no clear
idiomatic way to route these sub-views in a unified manner.

This prevents page reload by using HTML5 pushState underneath.